Fixed timecode warnings and opus audio decoding.
Former-commit-id: 4b2f0f30ebbf57035c8b859da05c982569cf75f1 [formerly 4b2f0f30ebbf57035c8b859da05c982569cf75f1 [formerly 14bac35a120125901bf30d66010f0a3a01ba044c]]
Former-commit-id: b1e07097efaece11ca43ee1dd139802e038807c2 [formerly d6a8a1d0d7505b8cb4c046e4c560e646ad252bf6]
Former-commit-id: 98c5b9d050fc9deb45d5f6f232250ca2a77acaf3

+++ mrViewer/src/gui/mrvTimecode.cpp
@@ -97,7 +97,7 @@ void Timecode::value( const int hours, const int mins, const int secs,
 		      const int frames )
 {
   int64_t x = frames;
-  int ifps = round(_fps);
+  int ifps = (int)round(_fps);
   x += secs * ifps;
   x += mins * 60 * ifps;
   x += hours * 3600 * ifps;
@@ -150,7 +150,7 @@ int64_t Timecode::value() const
                 assert( secs < 60 );
                 assert( mins < 60 );
 
-                int ifps = round(_fps);
+                int ifps = (int)round(_fps);
 
                 int64_t r = frames + 1;  // timecode starts at 0, frames at 1
                 r += secs * ifps;
@@ -323,7 +323,7 @@ int Timecode::format( char* buf, const mrv::Timecode::Display display,
                 int  frames = 0;
 
                 // Convert current frame value to timecode based on fps
-                int ifps = round(fps);
+                int ifps = (int)round(fps);
                 int frames_per_hour = 3600 * ifps;
                 int minbase  = 60 * ifps;
                 int secbase  = ifps;
@@ -331,13 +331,13 @@ int Timecode::format( char* buf, const mrv::Timecode::Display display,
 
                 // timecode starts at 0 so we substract 1 from frame
                 int64_t x = f - 1 + tc;
-                hours = x / frames_per_hour;
+                hours = int( x / frames_per_hour );
                 x -= hours * frames_per_hour;
-                mins = x / minbase;
+                mins = int( x / minbase );
                 x -= mins * minbase;
-                secs = x / secbase;
+                secs = int( x / secbase );
                 x -= secs * secbase;
-                frames = x;
+                frames = int(x);
 
                 // If negative timecode, make hour negative only
                 hours = abs(hours); mins = abs(mins); secs = abs(secs);
